Move three single-purpose forms from components/ to views/controllers/
Per .claude/rules/phlex_conversions.md, forms that only render for a
specific controller (or controller tree) belong under
`app/views/controllers/<controller>/form.rb`. The components/ folder
is reserved for genuinely reusable building blocks. Three forms were
violating that rule:

- `Components::DescriptionForm` →
  `Views::Controllers::Descriptions::Form`
  Top-level `descriptions/` namespace because both
  `Names::DescriptionsController` and `Locations::DescriptionsController`
  render it. The explicit form-id override
  (`name_description_form` / `location_description_form`) is
  preserved — the auto-derived `description_form` would lose the
  name-vs-location distinction the controllers' tabs and tests
  rely on.

- `Components::ExternalLinkForm` →
  `Views::Controllers::Observations::ExternalLinks::Form`
  Only rendered by `Observations::ExternalLinksController` (both
  inline and via `ModalTurboForm`, which auto-resolves from
  `controller_path` — no caller-side change needed there). Form id
  follows the views_controller_form_id derivation:
  `external_link_form` → `observation_external_link_form`. Updated
  the two `assert_select` calls in the controller test.

- `Components::OccurrenceProjectsForm` →
  `Views::Controllers::Occurrences::Projects::Form`
  Renders inside the auto-opening confirmation modal on the
  occurrence new/edit pages and field_slips new/edit pages, all
  posting to `Occurrences::ProjectsController#update` (or
  `OccurrencesController#create`). No external form-id references —
  the modal targets a separate `modal_identifier`.

Each form's component test moves with it
(`test/views/controllers/...`) and uses unqualified sibling
references (`Form.new(...)`) inside its `module` block.
