+++ app/views/controllers/locations/descriptions/new.html.erb
@@ -5,5 +5,6 @@ add_page_title(
 add_context_nav(location_description_form_new_tabs(description: @description))
 %>
 
-<%= render(Components::DescriptionForm.new(@description, licenses: @licenses,
-                                            user: @user)) %>
+<%= render(Views::Controllers::Descriptions::Form.new(
+             @description, licenses: @licenses, user: @user
+           )) %>
