+++ app/views/controllers/names/descriptions/new.html.erb
@@ -3,5 +3,6 @@ add_page_title(:create_name_description_title.t(name: @name.display_name))
 add_context_nav(name_description_form_new_tabs(description: @description))
 %>
 
-<%= render(Components::DescriptionForm.new(@description, licenses: @licenses,
-                                            user: @user)) %>
+<%= render(Views::Controllers::Descriptions::Form.new(
+             @description, licenses: @licenses, user: @user
+           )) %>
