+++ app/views/controllers/observations/external_links/new.html.erb
@@ -6,7 +6,7 @@ add_new_title(:add_object, :EXTERNAL_LINK)
 
 <div class="row">
   <div class="col-xs-12 col-sm-7">
-    <%= render(Components::ExternalLinkForm.new(
+    <%= render(Views::Controllers::Observations::ExternalLinks::Form.new(
           @external_link,
           observation: @observation,
           sites: @sites,
